+++ docs/history.rst
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+Next
+----
+- Remove ``Pool.start()`` in favor of starting the pool upon Pool
+  instantiation.
+- Rename ``Pool.stop()`` to ``Pool.join()``
+- Add ``Pool.close()`` to drain the queue and shutdown consumer processes in
+  the background.
+- Add ``Pool.terminate()`` to kill the consumer processes.
+- Refactor a few things.
+
 0.4.0 (2018-01-27)
 ------------------
 - Essentially an API rewrite on what consumers are and how they consume.